More data.
These files are from a dataset I found online. It has a total of 11430 rows with many columns. And the best part is that 50% of them are phishing links. So the file 'extra_phishing.csv' is the one with all the phishing links from the 'extra_data_lnks.csv' file. 

PS: - You might see a commit that was reverted after a couple of minutes. Don't worry about it. I actually committed it directly into the main branch instead creating a PR, so its my bad :sweat_smile: